requirements.txt
# ── Web framework ────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20        # Required for FastAPI file uploads

# ── LangChain ecosystem ──────────────────────────────────────────────────────
langchain==0.3.14
langchain-core==0.3.29
langchain-community==0.3.14
langchain-chroma==0.1.4
langchain-ollama==0.2.2

# ── Vector store ─────────────────────────────────────────────────────────────
chromadb==0.6.3

# ── PDF parsing ──────────────────────────────────────────────────────────────
docling==2.15.0
docling-core==2.10.0

# ── Image processing ─────────────────────────────────────────────────────────
Pillow==11.1.0

# ── HTTP client (Ollama API calls) ───────────────────────────────────────────
httpx==0.28.1

# ── Data handling ─────────────────────────────────────────────────────────────
pandas==2.2.3
tabulate==0.9.0                 # Required for DataFrame.to_markdown()

# ── Configuration ─────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Validation ───────────────────────────────────────────────────────────────
pydantic==2.10.5
